requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==1.24.3
pandas==2.1.0
scikit-learn==1.3.0
sentence-transformers==2.2.2
hdbscan==0.8.33
Werkzeug==3.0.0
python-dotenv==1.0.0
openai==1.6.1
httpx==0.27.2